Update according to feedbacks and to align more closely with WebGL.
We now call on the browser to properly fallback to a different compute device if the targeted device e.g. GPU is unable to support compilation of model with unsupported data type e.g. float16. Hence, we eliminate the need to define error code in such event. The browser is required to handle this failure internally and fallback as appropriate.

We also like to scope the current CompilationPreference to just the power options aka PowerPreference with two options available { 'low-power', 'speed' }. These are the same two options available in WebGLPowerPreference enum. The fact that we make the power preference option optional means we do not need to explicit define the 'default' option.