Don't toggle details when clicking on definition/variables
Bikeshed makes definitions (algorithms, methods) and variables
clickable and show valuable info. This collides with this spec's use
of details elements to allow showing/hiding the algorithms, where
clicking in the summary, by default, toggles the details.

During setup of the spec, slip in a "preventDefault()" to make these
play better together.

Details in #450