Wording change: Streamline context creation and validation algorithms
Bundle validation right into context creation, rather than making it
a separate step.

Note that with the removal of the "validate MLContext" algorithm, a
handful of the internal enum members are never referenced; they are
implicitly mapped to identical enums defined in Web IDL. This is
discussed in #497 which proposes de-duping the enums. In the mean
time, the definitions are marked with "ignore" to silence warnings
about unreferenced defnitions.

Fixes #498