Bug fix: Remove null value support in reshape()'s newShape parameter
As noted in issue #388, the reshape op build method currently supports
a special null value in newShape array that instructs the
implementation to compute the size of the dimension with the value
null so that the total size remains constant. @fwdr proposed WebNN
should only support the concrete values in newShape because this can
be easily handled by framework through shape inference

It appears this change has already landed in the Chromium
implementation and WPTs.

Fixes #388