Style: Make grouped algorithm steps consistent
For operations defined in bulk - element-wise binary, unary, and
logical ops, plus pooling and reduction ops - make the styling
consistent:

- A single <div class=algorithm-steps> around all definitions
- A separate <div algorithm> around each op's steps.

This was already done for the element-wise ops so already the majority
style. Pooling ops now have just a single "Algorithm" watermark
(instead of 3), and reduction ops now have a watermark instead of
none.

For #450