Bug fix: Drop "... have been checked..." notes
A smattering of methods on MLGraphBuilder (build, buildSync, concat,
constant, expand, and input) had a note about permissions and context
validity at the start of the steps. This is implicit, and applies to
all methods so it was confusing - was there something special about
these methods?

Nope. So remove the notes.

Fixes #501