Decompositions for reduceLogSum, reduceLogSumExp, and reduceSumSquare
As noted in #209, not all frameworks/platform APIs support all reduce
operations, but these are easily emulated.

Decompositions are provided using the standard phrasing and style
currently used throughout the spec. In the future, we may alter the
wording to explicitly state that these decompositions can be used as
templates for implementations to follow when their underlying
frameworks/platform APIs do not support the operations.

Fixes #209