Editorial: Make "generically emulated" text a macro, update wording (#638)
Many complex operations in the spec are given decompositions, showing
how the operation could be replaced through the use of more primitive
operations. This text is repeated in many places. Improve this in two
ways:

* Use Bikeshed's Text Macro[1] to reduce the repetition.

* Streamline the text, and make it explicit that if the underlying
  platform doesn't support an operation, WebNN API implementations can
  use the decomposition as a guide to emulate it.

The macro [EMULATED] is used at the end of an intro sentence since
there are variations - some of the decompositions are grouped, and
some make assumptions (e.g. activation functions, layouts, etc).

If we assume that implementations must implement all operations in the
spec, either via the underlying platform or emulation, this fixes #187

[1] https://speced.github.io/bikeshed/#metadata-text-macro