Revise graph resource validation (#622)
* Bug fix: Make "validate graph resources" test reflexive

compute() calls validate graph resources with the passed input/output
maps and the graph's input/output descriptors. The intent is to ensure
that all of a graph's descriptors are matched to a passed resource.
However, the steps were only validating that a passed resource had a
corresponding descriptor - it wouldn't flag if a resource was missing!

The Chromium prototype implementation makes the test reflexive - all
descriptors must have a corresponding resource and all resources must
have a corresponding descriptor. Incorporate that into the spec in the
same way - if number of keys is the same, and each resource has a
descriptor, then _ipso facto_* each descriptor has a resource.

Fixes #602

(* = a Latin phrase meaning "the writer is pretentious")

* Update per discussion in PR review

* inputs must be provided, but extra inputs are ignored
* requested outputs must exist, but not all outputs must be requested

---------

Co-authored-by: Dwayne Robinson <dwayner@microsoft.com>