Add axis argument to softmax()
Frameworks (TensorFlow, PyTorch, ONNX) all accept an axis parameter.

Most backends also support an axis, or it can be emulated with a
reshape. As @fdwr wrote: So it's achievable in each backend... but it
would move the pain from the caller down to where it can be handled
efficiently.

Fixes #466