Style: Consistently use backticks for expressions in prose
Outside of algorithms, the spec is mostly consistent about formatting
expressions using `backticks`; tidy up the few exceptions and add a
convention note.

Also, apply the existing convention to format tensor shape definitions
using *[ ... ]* consistently.

Also also, since it commonly precedes an expression, fix "as follow"
to be "as follows" - the latter is the correct English form for this
phrase, per style guides, even if multiple items follow it.