Use consistent phrasing for operator creation
Places that create ops use an inconsistent mixed of simple phrases,
camelCase, Title Case, ACRONYMS, and "quoted strings". The most common
was camelCase, but the wording can be weird, and the bulk-defined
binary/unary/logical/pooling/reduction ops and activations use a
"quotedCamelCase", so I went with that.

See https://github.com/webmachinelearning/webnn/pull/591#issuecomment-1970083381
for the most commentary.

Resolves #572