[Conventions] Avoid "sequence" in prose, linkify use in types
Per the spec processing model, algorithms don't usually deal with a
sequence<T> type; the JS value has been converted to an Infra list. So
try to avoid using that term in prose, especially as it is confusing
for recurrent operators which have temporal sequences as well. But do
call out an exception: unpacking unions.

When used in an IDL type description, though, do linkify it.

Fix a few other grammar/link glitches, too.