Add optional operator labels for more diagnosable error messages
This adds an internal 'label' property to the operators that are
created as a graph is constructed, which MAY (in the RFC 2119 sense)
be used by implementations in async error messages. Developers
populate this via a 'label' member in the options dictionary for
MLGraphBuilder methods.

A new MLOperatorOptions dictionary is defined, and all existing
options dictionaries now inherit from this, and all relevant methods
now take an options dictionary.

Fixes #585