Fix returning failure from methods, and use "item" for lists
Two glitches introduced in a88bb0e where validation steps were
inlined; the out-of-line steps returned "failure" but now that they're
inlined they need to explicitly throw. My bad.

Also, when referring to the contents of a list, the Infra term is
"items", so use that and linkify it, rather than the more generic
"elements" term. We more consistently use "elements" when referring to
the contents of a tensor anyway.