Fix harmless regex glitch
The regex included `(:? ... )` instead of `(?: ... )` when intending a
non-capturing group, but this was harmless as (1) the captured group
was ignored anyway and (2) the `:?` construct would helpfully match
nothing. The full group would end up being something like
`(:?dt|dd|/dl)` which is just slightly a superset of the intended
`(?:dt|dd|/dl)` and the capture is ignored anyway, so simplify the
regex.